Restore lean defaults for memory efficiency, fix test targets
Revert MAX_VARS to 100 and MAX_MEM_LISTS to 500 (from 500/2500) to
match LADR-2022-12A memory behavior. Add #ifndef guards so both can
be overridden via XFLAGS for fat builds. Revert MAX_BASIS to 100
(must be <= MAX_VARS). Shrink just struct ID fields from unsigned
long long to int. Fix Makefile test targets with -f flag and
< /dev/null to prevent stdin hangs.

Bump version to 2026-3J.

+++ ladr/just.h
@@ -64,8 +64,8 @@ typedef enum {
 typedef struct parajust * Parajust;
 
 struct parajust {
-  unsigned long long from_id;
-  unsigned long long into_id;
+  int from_id;
+  int into_id;
   Ilist from_pos;
   Ilist into_pos;
 };
@@ -73,7 +73,7 @@ struct parajust {
 typedef struct instancejust * Instancejust;
 
 struct instancejust {
-  unsigned long long parent_id;
+  int parent_id;
   Plist pairs;
 };
 
@@ -81,8 +81,8 @@ typedef struct ivyjust * Ivyjust;
 
 struct ivyjust {
   Just_type type;  /* input,resolve,paramod,instantiate,flip,propositional */
-  unsigned long long parent1;
-  unsigned long long parent2;
+  int parent1;
+  int parent2;
   Ilist pos1;
   Ilist pos2;
   Plist pairs;  /* for instantiate */
@@ -94,7 +94,7 @@ struct just {
   Just_type type;
   Just next;
   union {
-    unsigned long long id;
+    int id;
     Ilist lst;
     Parajust para;
     I3list demod;
